requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0 